shop_order_service/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
httpie==3.2.2
idna==3.7
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
Pygments==2.18.0
PyJWT==2.8.0
PySocks==1.7.1
pytest==8.2.1
pytest-django==4.8.0
pytz==2024.1
redis==5.0.5
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==70.0.0
sqlparse==0.5.0
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
